requirements.txt
Flask==2.3.3
qrcode==7.4.2
Pillow==10.0.1
mysql-connector-python==8.0.33
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.5
qrcode==7.3.1
mysql-connector-python==8.0.32
python-dotenv==0.21.0